fix: address PR review feedback (code quality and docs)
- Extract versionCode logic to AndroidPackageInfo helper utility
- Remove redundant versionName property, use appVersion directly
- Improve OTel service comments for mobile developer clarity
- Fix Markdown formatting in Android README (links, indentation, bold/code mixing)
- Update sourcemaps.config.example.json with safe placeholder values

All changes improve code readability and maintainability while preserving
the demo's role as a reference implementation for customers.
