+++ Mobiles/react-native/sourcemaps.config.example.json
@@ -1,7 +1,7 @@
 {
   "appName": "QuickPizza_ReactNative",
-  "endpoint": "http://localhost:8000/faro/api/v1",
-  "appId": "1",
-  "stackId": "1",
-  "apiKey": "auth_token_value"
+  "endpoint": "https://your-stack.grafana.net/api/v1",
+  "appId": "your-frontend-observability-app-id",
+  "stackId": "your-grafana-cloud-stack-id",
+  "apiKey": ""
 }
